Gave operator buttons classes, added operator logic
Gave buttons that represent operators specific classes to keep track of them and to specifically select them for JS use.
Subsequently added basic logic for how said operators will interact with the UI e.g. clearing the display and removing classes or only allowing one instance of a character to show on display.