Added per click button focus and bug fixes
Added logic for operator buttons that only adds the 'opActive' class to one of them at a time. Fixed a bug where the output from decimal button replaces the text in display if the only text in it is 0. Also, if the display content is an empty string, and the decimal button is clicked, it will add a zero (mainly aesthetics purposes as it doesn't affect mathematical operations if 0 is left out)